cloud config template: replace configmap
Replace the config map used to store cloud config templates with a
secret.

This change replaces the Kubernetes object used to store the cloud
config template data for newly created cloud config templates.
Additionially, it replaces the kind of Kubernetes object the various
lists and menus are based off of.
With this change, there is no longer a need for the Harvester UI to
receive and maintain a list of config maps from the Kubernetes API
server, which slightly speeds up the UI.

related-to: harvester/harvester#2329

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>