fix: vlan id input mask
Fix number validation logic in VLAN ID input mask in the VM network
creation/edit dialogue to allow input of any number between 1 and 4094,
as intended.
The new logic does away with regex and just checks if the input value is
number-like and within the allowed range. For any number-like outside
the range, the input is clamped to the closest value within the range.

related-to: SURE-8851

Signed-off-by: Moritz Röhrich <moritz.rohrich@suse.com>